+++ zcash_client_sqlite/src/lib.rs
@@ -609,6 +609,24 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         )
     }
 
+    #[cfg(feature = "transparent-inputs")]
+    fn get_spendable_transparent_outputs_for_addresses(
+        &self,
+        addresses: &[TransparentAddress],
+        target_height: TargetHeight,
+        confirmations_policy: ConfirmationsPolicy,
+        output_filter: TransparentOutputFilter,
+    ) -> Result<Vec<WalletTransparentOutput<Self::AccountId>>, Self::Error> {
+        wallet::transparent::get_spendable_transparent_outputs_for_addresses(
+            self.conn.borrow(),
+            &self.params,
+            addresses,
+            target_height,
+            confirmations_policy,
+            output_filter,
+        )
+    }
+
     /// Returns metadata for the spendable notes in the wallet.
     fn get_account_metadata(
         &self,
@@ -1399,6 +1417,15 @@ impl<C: BorrowMut<rusqlite::Connection>, P: consensus::Parameters, CL: Clock, R:
     ) -> Result<(), Self::Error> {
         self.transactionally(|wdb| wdb.notify_address_checked(request, as_of_height))
     }
+
+    #[cfg(feature = "spend-index")]
+    fn notify_output_verified_unspent(
+        &mut self,
+        outpoint: OutPoint,
+        as_of_height: BlockHeight,
+    ) -> Result<(), Self::Error> {
+        self.transactionally(|wdb| wdb.notify_output_verified_unspent(outpoint, as_of_height))
+    }
 }
 
 /// This impl block is only usable when you already have an [`SqlTransaction`], meaning
@@ -1796,6 +1823,19 @@ impl<P: consensus::Parameters, CL: Clock, R: RngCore> WalletWrite
             self.conn.0,
             &self.params,
             request.address(),
+            as_of_height,
+        )
+    }
+
+    #[cfg(feature = "spend-index")]
+    fn notify_output_verified_unspent(
+        &mut self,
+        outpoint: OutPoint,
+        as_of_height: BlockHeight,
+    ) -> Result<(), Self::Error> {
+        wallet::transparent::update_observed_unspent_height_for_outpoint(
+            self.conn.0,
+            &outpoint,
             as_of_height,
         )
     }